update mkiso.sh
this commit will bring a series of improvements and bug fixes!
now all of the functions of the script will work properly.

+++ mkiso.sh
@@ -9,33 +9,69 @@
 # 				 basta rodar o script especificando o profile que deseja compilar
 #				 exemplo: mkiso <profile>
 
-profile="$(./profile)"
-workdir="$(./work)"
-outdir="$(./iso)"
-# REGRA
+# VARS
 
-if [ -z "$profile"  ]; then
-	echo "você não pode passar um profile vazio."
-	exit 1
-fi
+profile="./profile"
+workdir="./work"
+outdir="./iso"
 
-if [ ! "$(id -u)" == 0  ]; then
-   echo "você precisa ser root para executar esse script"
-   exit 1
-fi
+# FUNC
+
+mkiso(){
+	mkarchiso -v -w "$workdir" -o "$outdir" "$profile"
+}
+
+verify_profile(){
+	if [ -z "$profile"  ]; then
+		printf '%s\n' "você não pode passar um profile vazio."
+		exit 1
+	fi
+}
+
+verify_dir(){
+	if [ -z "$1"  ]; then
+		printf '%s\n' "você não pode passar um profile vazio."
+		exit 1
+	fi
+}
+
+verify_root(){
+	if [ ! "$(id -u)" == 0  ]; then
+	   printf '%s\n' "você precisa ser root para executar esse script" 
+	   exit 1
+	fi
+}
+
+clean(){
+	rm -vrf "${workdir:?}"/*
+}
+
+verify_and_clean(){
+	if [ "$(ls -A $workdir)" ]; then
+	   printf '%s\n' "diretório não vazio, limpando..."
+	   clean 
+	fi
+}
+
+# VERIFICANDO ROOT
+
+verify_root
+
+# VERIFICANDO SE OS DIRETÓRIOS EXISTEM
+
+verify_profile
+verify_dir "$workdir"
+verify_dir "$outdir"
 
 # VERIFICANDO E LIMPANDO WORK DIR
 
-if [ "$(ls -A $workdir)" ]; then
-   echo "diretório não vazio, limpando..."
-   rm -vrf "$workdir"/*
-fi
+verify_and_clean
 
 # INICIANDO BUILD
 
-mkarchiso -v -w "$workdir" -o "$outdir" "$profile"
+mkiso
 
 # LIMPANDO WORKDIR APÓS O FINAL DO BUILD
 
-echo "limpando workdir"
-rm -vrf "$workdir"/*
+printf '%s\n' "limpando workdir"
+clean
